Change MediaStatusModel.id to Integer, add V22 migration, fix Javadoc
Agent-Logs-Url: https://github.com/EspacoGeek-Teams/SpringAPI_EspacoGeek/sessions/bbdd08fd-8299-45d3-9df2-a15df46d6d4d

+++ src/main/java/com/espacogeek/geek/repositories/MediaStatusRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.espacogeek.geek.models.MediaStatusModel;
 
 @Repository
-public interface MediaStatusRepository extends JpaRepository<MediaStatusModel, Long> {
+public interface MediaStatusRepository extends JpaRepository<MediaStatusModel, Integer> {
 
 }
